栈相关递归操作
stack reverse, stack_fifo_push/pop

+++ algorithm_playground/algorithm_playground/min_heap.cpp
@@ -19,7 +19,7 @@ DECLARE_MAIN_ENTRY(mh_unit_test_case);
 
 void mh_unit_test_case()
 {
-	srand( time(NULL) );
+	srand( (unsigned int)(time(NULL)) );
 	
 	const int kMaxSize = 51;
 	const int kMaxValue = 100;
